#40 Added edit function to comments

+++ backend/src/main/java/com/soen341/instagram/service/impl/CommentService.java
@@ -1,18 +1,23 @@
 package com.soen341.instagram.service.impl;
 
 import java.util.Date;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.soen341.instagram.dao.impl.AccountRepository;
 import com.soen341.instagram.dao.impl.CommentRepository;
 import com.soen341.instagram.dao.impl.PictureRepository;
+import com.soen341.instagram.dto.comment.CommentResponseDTO;
+import com.soen341.instagram.dto.picture.PictureDTO;
 import com.soen341.instagram.exception.comment.CommentLengthTooLongException;
 import com.soen341.instagram.exception.comment.CommentNotFoundException;
+import com.soen341.instagram.exception.comment.UnauthorizedRightsException;
 import com.soen341.instagram.exception.like.MultipleLikeException;
 import com.soen341.instagram.exception.picture.InvalidIdException;
 import com.soen341.instagram.exception.picture.PictureNotFoundException;
@@ -30,7 +35,8 @@ public class CommentService
 	private AccountRepository accountRepository;
 	@Autowired
 	private PictureRepository pictureRepository;
-
+	@Autowired
+	private ModelMapper modelMapper;
 	private static int maxCommentLength = 250;
 
 	public Comment createComment(final String commentContent, final long pictureId)
@@ -92,7 +98,14 @@ public class CommentService
 	public void deleteComment(final long commentId)
 	{
 		final Comment comment = findComment(commentId);
-		commentRepository.delete(comment);
+		if (comment.getAccount().getUsername().equals(UserAccessor.getCurrentAccount(accountRepository).getUsername()))
+		{
+			commentRepository.delete(comment);
+		}
+		else
+		{
+			throw new UnauthorizedRightsException();
+		}
 	}
 
 	public Comment editComment(final long commentId, final String newComment)
@@ -103,8 +116,17 @@ public class CommentService
 		}
 
 		final Comment comment = findComment(commentId);
-		comment.setComment(newComment);
-		commentRepository.save(comment);
+
+		if (comment.getAccount().getUsername().equals(UserAccessor.getCurrentAccount(accountRepository).getUsername()))
+		{
+			comment.setComment(newComment);
+			commentRepository.save(comment);
+		}
+		else
+		{
+			throw new UnauthorizedRightsException();
+		}
+
 		return comment;
 	}
 
@@ -132,5 +154,59 @@ public class CommentService
 			throw new CommentNotFoundException();
 		}
 		return commentOptional.get();
+	}
+
+	// Determine whether the current user can edit a comment
+	public List<CommentResponseDTO> determineEditable(final List<Comment> comments)
+	{
+		final List<CommentResponseDTO> commentsResponseDTO = new LinkedList<CommentResponseDTO>();
+
+		if (comments.isEmpty())
+		{
+			return commentsResponseDTO;
+		}
+
+		final Account currentUserRequest = UserAccessor.getCurrentAccount(accountRepository);
+		final String currentUser = currentUserRequest.getUsername();
+
+		for (final Comment comment : comments)
+		{
+			commentsResponseDTO.add(convertCommentIntoDTOWithUser(comment, currentUser));
+		}
+
+		// If the picture belongs to the current user, he can delete any comments
+		if (comments.get(0).getPicture().getAccount().getUsername().equals(currentUser))
+		{
+			for (final CommentResponseDTO commentResponseDTO : commentsResponseDTO)
+			{
+				commentResponseDTO.setEditable(true);
+			}
+		}
+
+		return commentsResponseDTO;
+	}
+
+	private CommentResponseDTO convertCommentIntoDTOWithUser(final Comment comment, final String currentUser)
+	{
+		final CommentResponseDTO commentResponseDTO = modelMapper.map(comment, CommentResponseDTO.class);
+		commentResponseDTO.setNbLikes(comment.getLikedBy().size());
+		// Creating picture DTO
+		final Picture picture = comment.getPicture();
+		final PictureDTO pictureDTO = modelMapper.map(picture, PictureDTO.class);
+		pictureDTO.setAccount(picture.getAccount().getUsername());
+
+		commentResponseDTO.setPictureDTO(pictureDTO);
+		commentResponseDTO.setAccount(comment.getAccount().getUsername());
+
+		if (comment.getAccount().getUsername().equals(currentUser))
+		{
+			commentResponseDTO.setEditable(true);
+		}
+		else
+		{
+			commentResponseDTO.setEditable(false);
+		}
+
+		return commentResponseDTO;
 	}
 }
